requirements.txt
google-genai
opencv-python
pillow
python-dotenv
pyttsx3
SpeechRecognition
sounddevice
scipy
numpy